Added UEFI support and FS stub.

+++ source/libk/include/string.h
@@ -8,7 +8,7 @@ extern "C" {
 #include <stddef.h>
 #include <stdint.h>
 
-int memcmp(const void*, const void*, size_t);
+size_t memcmp(const void*, const void*, size_t);
 void *memcpy(void*, void*, size_t);
 void *memset(void*, uint8_t, size_t);
 void *memmove(void*, void*, size_t);
