+++ source/libk/string/memcmp.c
@@ -3,7 +3,7 @@
 
 
 
-int memcmp(const void *aptr, const void *bptr, size_t size) {
+size_t memcmp(const void *aptr, const void *bptr, size_t size) {
 	const uint8_t *a = (const uint8_t *) aptr;
 	const uint8_t *b = (const uint8_t *) bptr;
